Prediction of the Cryptocurrency Behavior Using Sentiment Analysis of Twitter and News data/CODE/frontend/requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
dash==2.3.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.1.1
Flask-Compress==1.11
fonttools==4.32.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
multitasking==0.0.10
newsapi-python==0.2.6
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
plotly==5.7.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
tenacity==8.0.1
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.64.0
urllib3==1.26.9
Werkzeug==2.1.1
wordcloud==1.8.1
yfinance==0.1.70
zipp==3.8.0